jit and knownJit routers
