+++ prisma/schema.prisma
@@ -1,5 +1,4 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
+// Useful Prisma Schema visualizer https://prisma-editor.vercel.app
 
 generator client {
   provider = "prisma-client-js"
@@ -47,12 +46,14 @@ model Jit {
 
   // Relationships
   users        User[]         @relation("JitParticipants")
-  positions    Position[]     @relation("PositionJits")
+  position     Position       @relation("PositionJits", fields: [positionId], references: [id])
   SequenceJits SequenceJits[]
   KnownJit     KnownJit[]
+  positionId   String         @default("")
 
   // Indexes
   @@index([category, percentage])
+  @@index([positionId])
 }
 
 enum UserRole {
